Moved downloader to a separate package
Fix: Invalid localization for protocols & blacklist
