+++ src/downloader/settings.ts
